draggable testing

+++ src/Configurations/src/Configs/View.java
@@ -1,5 +1,8 @@
 package Configs;
 
+import com.thoughtworks.xstream.annotations.XStreamOmitField;
+
+
 public class View implements Configurable {
     private Configuration myConfiguration;
 
@@ -10,9 +13,9 @@ public class View implements Configurable {
 //    @Configure
 //    private String imageName;
     @Configure
-    private int width;
+    private double width;
     @Configure
-    private int height;
+    private double height;
     private Configurable myConfigurable;
 
 
@@ -22,7 +25,7 @@ public class View implements Configurable {
     }
 
     //this constructor is for the special case for the terrain blocks in the map
-    public View(String file, int width, int height) {
+    public View(String file, double width, double height) {
         imagePath = file;
         this.width = width;
         this.height = height;
@@ -53,10 +56,10 @@ public class View implements Configurable {
 //        myImageView.getFitWidth();
 //    }
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
     public String getImage() {
